requirements.txt
flask>=2.0.0
authlib>=1.2.0
python-dotenv>=1.0.0